release: v0.12.13 品質・性能強化リリース
- src/cli/workflow.rs (score=1197) を common/check/fmt/config_cmd サブモジュールに分割
- md059.rs::normalize_link_text の中間 Vec アロケーションを除去
- Windows CI パスセパレータ互換性修正（ast_linter テスト）
- coverage-baseline を 873 → 880 に更新（分割による行数増加）
- OpenSpec v0-12-13 をアーカイブ
